Move LoggerConstants.kt to the 'warehouse' package and fix imports.
