+++ src/main/kotlin/hu/netsurf/erp/warehouse/controller/ProductPhotoController.kt
@@ -1,8 +1,10 @@
 package hu.netsurf.erp.warehouse.controller
 
-import hu.netsurf.erp.common.logging.constants.LoggerConstants.PHOTO_FILE_NAME
-import hu.netsurf.erp.common.logging.constants.LoggerConstants.PRODUCT_ID
-import hu.netsurf.erp.common.logging.constants.warehouse.LogEventConstants
+import hu.netsurf.erp.common.logging.constants.warehouse.LogEventConstants.UPLOAD_PRODUCT_PHOTO_FAILURE_RESPONSE
+import hu.netsurf.erp.common.logging.constants.warehouse.LogEventConstants.UPLOAD_PRODUCT_PHOTO_REQUEST_RECEIVED
+import hu.netsurf.erp.common.logging.constants.warehouse.LogEventConstants.UPLOAD_PRODUCT_PHOTO_SUCCESS_RESPONSE
+import hu.netsurf.erp.common.logging.constants.warehouse.LoggerConstants.PHOTO_FILE_NAME
+import hu.netsurf.erp.common.logging.constants.warehouse.LoggerConstants.PRODUCT_ID
 import hu.netsurf.erp.common.logging.extension.logError
 import hu.netsurf.erp.common.logging.extension.logInfo
 import hu.netsurf.erp.warehouse.constants.EndpointConstants.CONTROLLER_PATH_PRODUCT_PHOTOS
@@ -33,14 +35,14 @@ class ProductPhotoController(private val productPhotoService: ProductPhotoServic
     ): ResponseEntity<String> {
         try {
             logger.logInfo(
-                LogEventConstants.UPLOAD_PRODUCT_PHOTO_REQUEST_RECEIVED,
+                UPLOAD_PRODUCT_PHOTO_REQUEST_RECEIVED,
                 mapOf(PRODUCT_ID to productId),
             )
 
             val photoFileName = productPhotoService.uploadPhoto(productId, file)
 
             logger.logInfo(
-                LogEventConstants.UPLOAD_PRODUCT_PHOTO_SUCCESS_RESPONSE,
+                UPLOAD_PRODUCT_PHOTO_SUCCESS_RESPONSE,
                 mapOf(
                     PRODUCT_ID to productId,
                     PHOTO_FILE_NAME to photoFileName.toString(),
@@ -50,14 +52,14 @@ class ProductPhotoController(private val productPhotoService: ProductPhotoServic
             return ResponseEntity(photoFileName, HttpStatus.OK)
         } catch (exception: NotFoundException) {
             logger.logError(
-                LogEventConstants.UPLOAD_PRODUCT_PHOTO_FAILURE_RESPONSE,
+                UPLOAD_PRODUCT_PHOTO_FAILURE_RESPONSE,
                 exception,
             )
 
             return ResponseEntity(exception.message, HttpStatus.NOT_FOUND)
         } catch (exception: Exception) {
             logger.logError(
-                LogEventConstants.UPLOAD_PRODUCT_PHOTO_FAILURE_RESPONSE,
+                UPLOAD_PRODUCT_PHOTO_FAILURE_RESPONSE,
                 exception,
             )
